ruby: Delete bundler.rb to avoid conflict with gem
We already delete the `bundler` directory but not the associated `bundler.rb` file. This causes a conflict with the separate rubygem-bundler package. With the file present, you need to call `gem "bundler"` before you can call `require "bundler"`. This change does the same thing that the Gentoo Linux ruby package does.